Move Wall class and labyrinth function to maze.py
